backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.30
aiosqlite==0.20.0
psutil==5.9.8
aiofiles==23.2.1
python-dotenv==1.0.1
websockets==12.0
httpx==0.27.0
pydantic[email]==2.7.1
redis[asyncio]==5.0.4
prometheus-client==0.20.0
